+++ config/config.exs
@@ -8,7 +8,8 @@ config :logger, :default_formatter,
   colors: [
     enabled: true,
     debug: :cyan,
-    info: :green,
+    info: :normal,
+    notice: :green,
     warning: :yellow,
     error: :red
   ]